Feat/#5: approverId, requestId, requesterId를 int64로 변경

+++ approval-processing-service/src/main/kotlin/org/dongchyeon/approvalprocessingservice/approval/grpc/ApprovalResultGrpcClient.kt
@@ -24,9 +24,9 @@ class ApprovalResultGrpcClient(
 
     fun sendResult(result: ApprovalResultPayload) {
         val request = ApprovalOuterClass.ApprovalResultRequest.newBuilder()
-            .setRequestId(result.requestId.toInt())
+            .setRequestId(result.requestId)
             .setStep(result.step)
-            .setApproverId(result.approverId.toInt())
+            .setApproverId(result.approverId)
             .setStatus(result.status.value)
             .build()
         blockingStub
